+++ app/src/main/java/com/example/cjj/data/AuthViewModelFactory.kt
@@ -2,7 +2,7 @@ package com.example.emptyactivity.data
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.emptyactivity.CJJApp
+import com.example.cjj.CJJApp
 
 class AuthViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
